backend/requirements/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
django-filter==2.4.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
gunicorn==20.1.0
isort==5.13.2